refactor: 입력받은 값 검증하고 이후 수행하는 로직 리팩토링 - 입력값 유효검사와 그에 따른 로직 수행을 checkResult 메서드 내부로 옮김 - play 메서드보다는 checkResult에서 로직을 수행하는 게 가독성 있어보이기 때문에 진행 - resetValue 메서드는 공용으로 사용되기 때문에 utils 폴더로 이동
